Add input delay to rectangular network
For some uses of rectangular networks (e.g. matrix multiplication)
it is important to be able to step the inputs of cells such that
the index of a cell along a dimension determines the number of
registers to place before the actual input of the cell.

This is implemented with a type that allows the two inputs to be
delayed independently. I can't think of an immediate use for this,
but it is trivial enough to implement so I see no downside.