Add input / output delay to `Rect`
The predefined rectangular network now optionally allows data
entering or leaving the network to be delayed. Output delay is
applied in a reverse order to input delay, meaning if both are
delayed there is effectively no delay (as values still enter and
leave at the same time), but the latency will increase accordingly.