Describe Ripple Carry Adder in README.md
There is now a working example of a ripple carry adder in the
README. While the code is complete and working, it would be nice
to include a diagram of the network architecture to make things
clearer. I should probably also include some ghci prompt
demonstration of the adder at the end of the example.