Add README and LICENSE (#4)
The project now has a README, which goes a long way to providing documentation which is useful for end users. This is not the main effort for documentation though (see issue #1).

The project has been licensed under the BSD 2 Clause License, and the license has been added to the project and all haddock module headers.